adk-backend/requirements.txt
# Google Agent Development Kit
adk-python==1.3.0

# Web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Google Cloud
google-cloud-aiplatform==1.38.1
google-auth==2.23.4

# Utilities
pydantic==2.5.0
python-dotenv==1.0.0
httpx==0.25.2

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
